+++ frontend/src/school/components/notice/Notice.jsx
@@ -1,6 +1,6 @@
 import { Box, Button, FormControl, InputLabel, MenuItem, Paper, Select, Skeleton, TextField, Typography } from '@mui/material'
 import { useFormik } from 'formik'
-import React, { useEffect, useState } from 'react'
+import React, { useEffect, useRef, useState } from 'react'
 import { noticeSchema } from '../../../yupSchema/noticeSchema.js'
 import axios from 'axios'
 import { baseApi } from '../../../environment.js'
@@ -10,15 +10,17 @@ import MessageSnackbar from '../../../basicUtilityComponent/snackbar/MessageSnac
 
 export default function Notice() {
   const [loading, setLoading] = React.useState(true);
+  const [notices, setNotices] = useState([]);
+  const [edit, setEdit] = useState(false);
+  const [editId, setEditId] = useState(null);
   const [message, setMessage] = React.useState('');
   const [messageType, setMessageType] = React.useState('');
   const handleMessageClose = () => {
     setMessage('');
   }
 
-  const [notices, setNotices] = useState([]);
-  const [edit, setEdit] = useState(false);
-  const [editId, setEditId] = useState(null);
+  const formRef = useRef(null);
+  const InputRef = useRef(null);
 
   const handleEdit = (x) => {
     setEdit(true);
@@ -26,6 +28,11 @@ export default function Notice() {
     formik.setFieldValue("title", x.title);
     formik.setFieldValue("message", x.message);
     formik.setFieldValue("audience", x.audience);
+
+    formRef.current?.scrollIntoView({ behavior: 'smooth', block: 'center' });
+    setTimeout(() => {
+      InputRef.current?.focus(); // Focus on the first field
+    }, 500);
   }
 
   const cancelEdit = () => {
@@ -35,15 +42,20 @@ export default function Notice() {
   }
 
   const handleDelete = (x) => {
-    axios.delete(`${baseApi}/notice/delete/${x._id}`)
-      .then(res => {
-        setMessage(res.data.message);
-        setMessageType("success");
-      }).catch(err => {
-        console.log("Error in delete Notices", err)
-        setMessage("Error in Delete");
-        setMessageType("error");
-      })
+    if (confirm("Are you sure! You want to Delete.")) {
+      axios.delete(`${baseApi}/notice/delete/${x._id}`)
+        .then(res => {
+          setMessage(res.data.message);
+          setMessageType("success");
+        }).catch(error => {
+          console.error(
+            `%c[ERROR in Deleting Notice]:- ${error.name || "Unknown Error"} `,
+            "color: red; font-weight: bold; font-size: 14px;", error
+          );
+          setMessage("Error in Delete");
+          setMessageType("error");
+        })
+    }
   }
 
   const formik = useFormik({
@@ -58,8 +70,11 @@ export default function Notice() {
             setMessage(res.data.message);
             setMessageType("success");
             cancelEdit();
-          }).catch(err => {
-            console.log("Error in Notice Updating:- ", err);
+          }).catch(error => {
+            console.error(
+              `%c[ERROR in Updating Notice]:- ${error.name || "Unknown Error"} `,
+              "color: red; font-weight: bold; font-size: 14px;", error
+            );
             setMessage("Error in update.");
             setMessageType("error");
           })
@@ -69,8 +84,11 @@ export default function Notice() {
             // console.log("Notice add response:- ", res);
             setMessage(res.data.message);
             setMessageType("success");
-          }).catch(err => {
-            console.log("Error in Notice:- ", err);
+          }).catch(error => {
+            console.error(
+              `%c[ERROR in Adding New Notice]:- ${error.name || "Unknown Error"} `,
+              "color: red; font-weight: bold; font-size: 14px;", error
+            );
             setMessage("Error in saving Notice.");
             setMessageType("error");
           })
@@ -85,8 +103,11 @@ export default function Notice() {
         setNotices(res.data.data);
         setLoading(false);
         // console.log(res)
-      }).catch(err => {
-        console.log("Error in fetching all Notices", err)
+      }).catch(error => {
+        console.error(
+          `%c[ERROR in Fetching Notices]:- ${error.name || "Unknown Error"} `,
+          "color: red; font-weight: bold; font-size: 14px;", error
+        );
         setLoading(false);
       })
   }
@@ -96,70 +117,73 @@ export default function Notice() {
 
   return (
     <>
-      <Typography variant='h3' sx={{ textAlign: 'center', fontWeight: '700' }}>Notice</Typography>
+      <Typography variant='h4' sx={{ fontWeight: '700' }}>Notice</Typography>
       {message && <MessageSnackbar message={message} messageType={messageType} handleClose={handleMessageClose} />}
-      <Box
-        component="form"
-        sx={{ '& > :not(style)': { m: 1 }, display: 'flex', flexDirection: 'column', width: '50vw', minWidth: '230px', margin: 'auto', background: '#fff' }}
-        noValidate
-        autoComplete="off"
-        onSubmit={formik.handleSubmit}
-      >
+      <Paper sx={{ marginBottom: 3}} ref={formRef}>
+        <Box
+          component="form"
+          sx={{ '& > :not(style)': { m: 1 }, display: 'flex', flexDirection: 'column', width: '50vw', minWidth: '230px', margin: 'auto', padding: 3 }}
+          noValidate
+          autoComplete="off"
+          onSubmit={formik.handleSubmit}
+        >
 
-        {edit ? <Typography variant='h4' sx={{ textAlign: "center", paddingBottom: "25px", fontWeight: 700 }}>Edit Notice</Typography>
-          : <Typography variant='h4' sx={{ textAlign: "center", paddingBottom: "25px", fontWeight: 700 }}>Add New Notice</Typography>
-        }
-        <TextField
-          name="title"
-          label="Title"
-          value={formik.values.title}
-          onChange={formik.handleChange}
-          onBlur={formik.handleBlur}
-        />
-        {formik.touched.title && formik.errors.title && <p style={{ color: 'red', textTransform: 'capitalize' }}>
-          {formik.errors.title}
-        </p>}
-
-        <TextField
-          multiline
-          rows={4}
-          name="message"
-          label="Message"
-          value={formik.values.message}
-          onChange={formik.handleChange}
-          onBlur={formik.handleBlur}
-        />
-        {formik.touched.message && formik.errors.message && <p style={{ color: 'red', textTransform: 'capitalize' }}>
-          {formik.errors.message}
-        </p>}
-
-        <FormControl sx={{ mt: '10px' }}>
-          <InputLabel id="demo-simple-select-label">Audience</InputLabel>
-          <Select
-            label="Audience"
-            name='audience'
-            value={formik.values.audience}
+          {edit ? <Typography variant='h4' sx={{ textAlign: "center", paddingBottom: "25px", fontWeight: 700 }}>Edit Notice</Typography>
+            : <Typography variant='h4' sx={{ textAlign: "center", paddingBottom: "25px", fontWeight: 700 }}>Add New Notice</Typography>
+          }
+          <TextField
+            name="title"
+            label="Title"
+            value={formik.values.title}
             onChange={formik.handleChange}
             onBlur={formik.handleBlur}
-          >
-            <MenuItem value="">Select Audience</MenuItem>
-            <MenuItem value="teacher">Teacher</MenuItem>
-            <MenuItem value="student">Student</MenuItem>
-          </Select>
-        </FormControl>
-        {formik.touched.audience && formik.errors.audience && <p style={{ color: 'red', textTransform: 'capitalize' }}>
-          {formik.errors.audience}
-        </p>}
+            inputRef={InputRef}
+          />
+          {formik.touched.title && formik.errors.title && <p style={{ color: 'red', textTransform: 'capitalize' }}>
+            {formik.errors.title}
+          </p>}
 
-        <Button sx={{ width: '120px' }} type='submit' variant='contained'>Submit</Button>
-        {edit && <Button sx={{ width: '120px' }} onClick={() => { cancelEdit() }} type='button' variant='outlined'>Cancel</Button>}
+          <TextField
+            multiline
+            rows={4}
+            name="message"
+            label="Message"
+            value={formik.values.message}
+            onChange={formik.handleChange}
+            onBlur={formik.handleBlur}
+          />
+          {formik.touched.message && formik.errors.message && <p style={{ color: 'red', textTransform: 'capitalize' }}>
+            {formik.errors.message}
+          </p>}
 
-      </Box>
+          <FormControl sx={{ mt: '10px' }}>
+            <InputLabel id="demo-simple-select-label">Audience</InputLabel>
+            <Select
+              label="Audience"
+              name='audience'
+              value={formik.values.audience}
+              onChange={formik.handleChange}
+              onBlur={formik.handleBlur}
+            >
+              <MenuItem value="">Select Audience</MenuItem>
+              <MenuItem value="teacher">Teacher</MenuItem>
+              <MenuItem value="student">Student</MenuItem>
+            </Select>
+          </FormControl>
+          {formik.touched.audience && formik.errors.audience && <p style={{ color: 'red', textTransform: 'capitalize' }}>
+            {formik.errors.audience}
+          </p>}
+
+          <Button type='submit' variant='contained'>Submit</Button>
+          {edit && <Button onClick={() => { cancelEdit() }} type='button' variant='outlined'>Cancel</Button>}
+
+        </Box>
+      </Paper>
 
       <Box component={'div'} sx={{ display: 'flex', flexDirection: 'row', flexWrap: 'wrap' }}>
         {loading ? (
           Array.from(new Array(5)).map((_, index) => (
-            <Paper key={index} sx={{ m: 2, p: 2, width: 500 }}>
+            <Paper key={index} sx={{ m: 2, p: 2, width: 400 }}>
               <Skeleton variant="text" width={200} height={60} />
               <Skeleton variant="text" width={250} height={40} />
               <Skeleton variant="text" width={180} height={40} />
